moved nnet read/write into module, tests passing OK
